Seperate class for buffer

+++ src/Pipeline.cpp
@@ -45,7 +45,7 @@ void Pipeline::createPipelineLayout(const PipelineParams& params)
     // Create the pipeline layout with the descriptor set layout and push constant range
     VkPipelineLayoutCreateInfo pipelineLayoutInfo{};
     pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-    pipelineLayoutInfo.setLayoutCount = params.descriptorSetLayouts.size();
+    pipelineLayoutInfo.setLayoutCount = static_cast<uint32_t>(params.descriptorSetLayouts.size());
     pipelineLayoutInfo.pSetLayouts = params.descriptorSetLayouts.data(); // Use the descriptor set layout from the params
     pipelineLayoutInfo.pushConstantRangeCount = 1;
     pipelineLayoutInfo.pPushConstantRanges = &params.pushConstantRange; // Use the push constant range from the params
